Use last page link to get total page count
For large issue counts, the pager only shows a limited number of links to
the individual pages, so just counting those links will give a capped
value. The link to the last page of issues is always present, and
extracting the page number from the url should always give the correct
page count.